+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/http/compiler/CompilerPluginTest.java
@@ -134,11 +134,14 @@ public class CompilerPluginTest {
         DiagnosticResult diagnosticResult = compilation.diagnosticResult();
         Assert.assertEquals(diagnosticResult.errorCount(), 3);
         assertError(diagnosticResult, 0, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'error[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'error[]'", HTTP_102);
         assertError(diagnosticResult, 1, "invalid resource method return type: expected 'anydata|http:Response" +
-                "|http:StatusCodeResponse|error', but found 'map<http:Client>'", HTTP_102);
+                "|http:StatusCodeResponse|stream<http:SseEvent, error?>|stream<http:SseEvent, error>|" +
+                "error', but found 'map<http:Client>'", HTTP_102);
         assertError(diagnosticResult, 2, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'readonly & error[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'readonly & error[]'", HTTP_102);
     }
 
     @Test
@@ -637,23 +640,32 @@ public class CompilerPluginTest {
         DiagnosticResult diagnosticResult = compilation.diagnosticResult();
         Assert.assertEquals(diagnosticResult.errorCount(), 9);
         assertTrue(diagnosticResult, 0, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord1[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord1[]'", HTTP_102);
         assertTrue(diagnosticResult, 1, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord2[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord2[]'", HTTP_102);
         assertTrue(diagnosticResult, 2, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord3[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord3[]'", HTTP_102);
         assertTrue(diagnosticResult, 3, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord4[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord4[]'", HTTP_102);
         assertTrue(diagnosticResult, 4, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord5[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>" +
+                "|stream<http:SseEvent, error>|error', but found 'TestRecord5[]'", HTTP_102);
         assertTrue(diagnosticResult, 5, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord6[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord6[]'", HTTP_102);
         assertTrue(diagnosticResult, 6, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'TestRecord7[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'TestRecord7[]'", HTTP_102);
         assertTrue(diagnosticResult, 7, "invalid resource method return type: expected 'anydata|" +
-                "http:Response|http:StatusCodeResponse|error', but found 'http:StatusCodeResponse[]'", HTTP_102);
+                "http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'http:StatusCodeResponse[]'", HTTP_102);
         assertTrue(diagnosticResult, 8, "invalid resource method return type: expected " +
-                "'anydata|http:Response|http:StatusCodeResponse|error', but found 'error[]'", HTTP_102);
+                "'anydata|http:Response|http:StatusCodeResponse|stream<http:SseEvent, error?>|" +
+                "stream<http:SseEvent, error>|error', but found 'error[]'", HTTP_102);
     }
 
     @Test