+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/Constants.java
@@ -90,7 +90,8 @@ public final class Constants {
     public static final String CALLER_ANNOTATION_NAME = "CallerInfo";
     public static final String FIELD_RESPONSE_TYPE = "respondType";
     public static final String RESPOND_METHOD_NAME = "respond";
-    public static final String ALLOWED_RETURN_UNION = "anydata|http:Response|http:StatusCodeResponse|error";
+    public static final String ALLOWED_RETURN_UNION = "anydata|http:Response|http:StatusCodeResponse|" +
+            "stream<http:SseEvent, error?>|stream<http:SseEvent, error>|error";
     public static final String REQUEST_INTERCEPTOR = "RequestInterceptor";
     public static final String RESPONSE_INTERCEPTOR = "ResponseInterceptor";
     public static final String REQUEST_ERROR_INTERCEPTOR = "RequestErrorInterceptor";